fota-demo: 96b_hikey: improve first boot docs
The first HiKey boot takes a long time, which is confusing and leads
people to believe the boot failed. Add a note about this so people
won't be surprised.

Also add a warning to remove the USB cable used to flash the device
when connecting to the network via USB Ethernet dongle.

Signed-off-by: Marti Bolivar <marti.bolivar@linaro.org>
Reported-by: Yang Zhang <yang.zhang@linaro.org>