Add scripts/get-mergeup-highlights.py
This is a script that can be run on repositories with upstreams. It
collects and prints the 'Higlights' content in each of the 'mergeup'
commits, per the Genesis branching rules described in
source/genesis/branching.rst.

This is useful for looking back at what happened in upstreams when
updating release notes, etc. The default indentation was chosen to
make copy/pasting into release-notes.yml convenient.

Signed-off-by: Marti Bolivar <marti.bolivar@linaro.org>