big: getting-started: restore whitelist documentation
The whitelist documentation was copied over incompletely. Restore the
old content, which contains complete usage instructions, slightly
modified to make it clear which commands are run in the container.

This is a bit hard to use, since running commands within the container
requires extra Docker knowledge, but this can be improved in the
future.

Signed-off-by: Marti Bolivar <marti.bolivar@linaro.org>