Restructure and clean up indexes to better fit microPlatforms.
Touch up the top-level indexes to correspond better to the OSF
structure.

The "basic IOT gateway" section makes less sense now that
gateway-containers is part of the Linux microPlatform, so move it into
linux/.

Delete top-level references to LTD, replacing them with OSF.

Clean up some of the language so the indexes flow better. This
includes moving a Zephyr-specific TODO into that mP's index.rst, and
touching up the formatting for when the todolist is shown.

Signed-off-by: Marti Bolivar <marti@opensourcefoundries.com>