linux: overhaul for OSF
Overhaul the source/linux directory to match the current state of play
for the OSF Linux microPlatform.

Rename basic-gateway.rst to iot-gateway.rst, because we don't want to
be basic. Clean up the presentation on installing Ansible on that
page, and make it clear that the latest version is required. Make a
clarifying point about this being a generic reference at the top of
the page, and point at the IoT foundry for step-by-step
instructions. Provide up to date information on the new repositories
and container registries.

The subscriber UX in iot-gateway.rst has some issues that could really
use some improvement, but it's accurate.

Remove the Ansible bits from the LMP getting-started page, since
that's moved to iot-gateway. Clean up and clarify other basic
information people will need as well, now that the scope is more
clear.

Signed-off-by: Marti Bolivar <marti@opensourcefoundries.com>