conf.py: support subscriber version from environment
Allow the environment to specify the subscriber version in the
OSF_UPDATE_SUBSCRIBER_VERSION variable. Fall back on git describe if
that's not available.

Note that due to the way sphinx-build works, this isn't particularly
robust -- a clean build is needed to change the value. That should be
OK for generating updates, since those builds are always clean, but it
makes developer building a little less convenient. Better options are
to be explored in the future.

Signed-off-by: Marti Bolivar <marti@opensourcefoundries.com>