Create new document tree structure
Add the following top-level directories with stubbed out index pages:

    - tutorial
    - howto
    - reference
    - other

These will contain the documentation in a restructured form. Content
from existing pages will be migrated into the new structure in
subsequent patches.

The rationale for this change is that the current documentation cannot
sensibly be read in any linear order, and it forces the reader to make
too many choices. Furthermore, recent investigations have shown that
the LWM2M demonstration system will be easier to secure first, leading
us to want to change emphasis there over the MQTT/hawkBit system.

The new documentation will include a tutorial that just describes the
LWM2M system on the recommended hardware, HOWTOs for changing that
system in ways we know users care about, more detailed reference
materials (from the current zephyr/ and linux/ directories), and
"other" content (where the MQTT system documentation will end up).

It will be possible to read it linearly, starting from the
tutorial. Users with particular goals in mind can then proceed to a
specific HOWTO if one exists. If not, or if more general information
is required, the reference manuals will be available. Random other
pages then logically follow last.

Signed-off-by: Marti Bolivar <marti@opensourcefoundries.com>