tutorial: installation-zephyr: update python dependencies
Add a new MCUboot dependency.

At some point we'd better start using a requirements.txt (or relying
on Zephyr's and MCUboot's, but Zephyr has the bad habit of specifying
explicit versions, so we might want to go down the rabbit hole of
setting up a virtualenv...). Let's not worry about that for now.

Signed-off-by: Marti Bolivar <marti@opensourcefoundries.com>