Add Wayback Machine support for running containers
We currently always ask users to run tip/latest containers. This
means documentation for old updates goes stale, which is a problem
that can only grow in severity.

Fix this issue by using the docker tag substitutions everywhere as
appropriate for the public and subscriber cases. Also fix up some of
the language that says "latest" and make some other wording changes to
keep things clean.

We still have a problem with source code checkouts: Git clones and
Repo inits all use the latest master. Fixing that is left to a
subsequent patch.

Signed-off-by: Marti Bolivar <marti@opensourcefoundries.com>