Add Wayback Machine support for Git and Repo.
Use the Git and Repo tag substitutions when cloning or initializing,
respectively, to ensure we can get the exact sources for this update.

This doesn't guarantee reproducible builds, but it's a step in the
right direction and at least allows users to rebuild the correct
source after the next update is released!

Signed-off-by: Marti Bolivar <marti@opensourcefoundries.com>