secure-boot: clarify the need to update the UUU scripts after fusing
Some projects might define a close.uuu script in their
meta-subscriber-overrride to perform further verification before
closing the device. It is unlikely that such a script will be
executing a signed SPL and therefore will not succeed as HAB events
will be triggering on boot.

This commit raises the awareness of the need to update all of the users UUU
scripts once the fuses have been burnt (before even closing the device).

Signed-off-by: Jorge Ramirez-Ortiz <jorge@foundries.io>