reference-manual: fix no-title caused by now superfluous index
The fact that we had an index for the reference manual has now been replaced by having it show up on the navbar to the left. To do this, we create a toctree in the root index of the docs. This means we have an index, which imports the old index which of course has no title, which causes the error we see.

Signed-off-by: Matthew Croughan <matthew.croughan@foundries.io>