replace package manager install guide with 'go get from source' instructions
Maintaining the packages is not going to work out since Andy and I are not communicating well. This needs to be automated, otherwise it won't work. I could do this, but I feel I do not have the permission to do so. The importance of this task is very low also.

Signed-off-by: Matthew Croughan <matthew.croughan@foundries.io>