getting-started::git-config use 'openssl base64' instead of 'base64' to mask platform differences
base64 on macOS does not have the -w0 option which usually disables linewrapping. Due to this, we would have had to have two different commands for linux/macos. Openssl works the same on Git Bash for Windows, and on macos/Linux, so I'm using it instead since it is more standard across platforms.

Signed-off-by: Matthew Croughan <matthew.croughan@foundries.io>