Add high-level explanation of QEMU CLI
Added a brief explanatory note of the QEMU CLI to start the image, and
included a link to QEMU's documentation. A more detailed explantation
could always be added later.

Checked that content appeared as intended when rendered as html.
Link was checked to ensure it was correct.

This addresses Jira FFTK-1035.

Signed-off-by: Katrina Prosise <katrina.prosise@foundries.io>