Expand on adding a systemd startup service
Added to the troubleshooting and FAQ page as a question/answer. The
initial plan was to move it onto a seperate tutorial page, but much of
the content was found to already be existing under the platform
customization page. Did some minor changes to the existing content, and
added a link to the tutorial as well as to the systemd site.

This commit applies to Jira FFTK #1068

Signed-off-by: Katrina Prosise <katrina.prosise@foundries.io>