Document use of the Docker credential helper
Added a page, `reference-manual/docker/configure-docker-helper`,
updated the sections index, and included links between api-access
page and the new page.

We may want to move it elsewhere, or change some of the content of the
new page.

Checked rendered pages locally for formatting, and ran the linkcheck.

This applies to Jira FFTK 1350

Signed-off-by: Katrina Prosise <katrina.prosise@foundries.io>