Cleanup getting-started install-fioctl page
Changes made to grammar, word usage, and admonition types. The sidebar
index page was modified so that install-fioctl came before the git page.
The git page assumes the user is familiar with generating a key,
which install-fioctl covers.

QA steps included running checklink and build, along with spellcheck.
The validity of the steps was not checked.

Signed-off-by: Katrina Prosise <katrina.prosise@foundries.io>