Add spldev key to RM:factory:factory-keys page
Added a minimal amount of info on spldev support. Decided against having
an admonition pointing out that there are two u-boot keys, which is
dependent on hardware features, to keep it short. Open to adding more
info if we think it is needed. There is cleanup tasks to be done
regarding word choice and grammar, but that will be in a following
PR.

Rendered page locally, and ran Vale linter. No *new* issues reported.

This addresses Jira FFTK-1597.

Signed-off-by: Katrina Prosise <katrina.prosise@foundries.io>