Clarify user permissions for Docker config helper
Two changes were made:
The use of sudo for the symlink step was changed to note that it is
specific to that cmd invocation specifically.
An additional admonition was added to draw attention to the need for
the user to belong to the docker group.

QA steps: make Linkcheck was ran, and the rendered html was checked.
Vale was ran to lint the content.

No related issue.

Signed-off-by: Katrina Prosise <katrina.prosise@foundries.io>