Add latest to canonical url path
The new path is https://docs.foundries.io/latest/.

Note that this only addresses the issue of going forward,
and does not completely solve the issue of older versions showing up in
search results.

Built and rendered locally without issue.

No issues to tag, as this is a relatively minor change. (picked out from
a larger effort)

Signed-off-by: Katrina Prosise <katrina.prosise@foundries.io>