Move team based access to new UG section
Moved the file from the `reference-manual/factory/` dir to
`user-guide/account-management/team-based-access`. A index page for the
new location was added, and the one for the reference manual was update.

To handle any potential links to
`.../latest/reference-manual/factory/team-based-access.html`, a redirect
was added, which required a Sphinx extension be added.

QA Steps: Tested the redirect locally with no issues, and checked the
rendering of the page, also without issues. It is possible that the
extension may need to have a version specified to build via the CI,
so the PR will also serve as part of the QA Steps.

This relates to Jira FFTK 1781

Signed-off-by: Katrina Prosise <katrina.prosise@foundries.io>