Add account management UI info to user guide
Added two pages to document account management tasks, covering
billing and access, to assist with common customer questions.
Team based access will be updated in a future commit to avoid
creep. The pages link to each other, which is why I made the decision to
bundle them in one commit.

An additional change was to move the account management section to the
top of the user guide, based on feedback.

Future changes will likely be needed as questions arise, and these pages
are not meant to be comprehensive at this time.

QA steps taken: quick pass with vale linter, site rendered as html and
visually inspected. Ran link-check. No issues were detected.

This commit applies to FFTK 1305 and FFTK 1611

Signed-off-by: Katrina Prosise <katrina.prosise@foundries.io>