Fix broken link reference in rm:security
Updated reference-manual/security/authentication-xilinx to
link to a listing of related documents, as the pdf we had been linking
to is no longer available at that address, and could not be found. The
information should be findable in the new link however.

Ran link-check and rendered the page. No errors were found.

This commit applies to Jira FFTK #1880

Signed-off-by: Katrina Prosise <katrina.prosise@foundries.io>