Add landing page for RM:FoundriesFactory section
Landing page added with a brief summary of FoundriesFactory.
Introductions to all the sections also listed in the page's toc was
added. A "seealso" admonition was added, linking to Factory account
management. Page labels were added to the page on SBOMs and for account
management.

QA steps include in buffer linting and checking the html rendering.

This commit applies to Jira FFTK 1903

Signed-off-by: Katrina Prosise <katrina.prosise@foundries.io>