Cleanup the flashing device getting started guide
Cleanup tasks focused on style rules and text cleanup. We may need
changes to content to bring it up to date. Screenshots show older GUI
colors, but should largely be the same.

QA Steps: Ran vale to lint, and then linkcheck and checked rendered
output. No issues should be present.

No Jira task to link to, as this is routine cleanup of the docs.

Signed-off-by: Katrina Prosise <katrina.prosise@foundries.io>