Add basic linter action for PRs
Adding the vale-action/reviewdog github action to run for PRs open
against main. Note that it will fail on errors, more work will be needed
to get suggestions and warnings posted automatically as review comments,
and it is still recommended that vale is run locally against changes
before opening a PR.

This was tested on a fork, so there may need to be changes (such as to
the TOKEN) before it works properly under the Foundriesio account.

Signed-off-by: Katrina Prosise <katrina.prosise@foundries.io>