Cleanup and split troubleshooting page
The troubleshooting page was split, with sections covering platform
customization getting their own page, which will later serve as
an umbrella for other sections in the user guide.

The troubleshooting page itself was moved from reference manual to the
user guide. Content was somewhat grouped and organized. There was a
number of formatting issues that were addressed. Further changes to
content may be needed.

Work was done to hopefully avoid any merge conflicts, as this commit has
been in progress for a good while.

Edits were also made to the content with updates and cleanup. Notably
the syntax for prepending in recipes was updated from

`FILESEXTRAPATHS_prepend := "${THISDIR}/${PN}:"`
to
`FILESEXTRAPATHS:prepend := "${THISDIR}/${PN}:"`

QA steps and results: checked rendered content and ran linkcheck. However `conf.py`
was changed to try and avoid merge issues, which made a final check not
possible. Redirects from troubleshooting page headers to new sections
under platform-customization are not functional; this means we may end
up with broken links, and Customer Service should be aware of this
before we decide to merge. Due to the breadth of subjects, not all
content could be verified for correctness in this cleanup, and would
benefit from multiple PR reviews.

No Jira issue to link.

Signed-off-by: Katrina Prosise <katrina.prosise@foundries.io>