Address existence of two board pages
Reference manual sections on flashing boards and supported boards were
not clear on what was covered. The flashing section was moved to the
user guide. A future issue can address the information present on the
supported board page to add more information, such as level of support;
this commit seeks to clarify the purposes of the two pages.

The support section currently links to flashing instructions, which is
fine for the time being.

The page on supported boards was also given a short admonition regarding
possible variability in support level.

The supported board page is now standalone under reference manual.

The csv file has also been related under
`reference-manual/supported-machines/` to make it easier to maintain.

ToC pages updated and Redirects added.

QA Steps: HTML checked, linted

This address Jira FFTK 991

Signed-off-by: Katrina Prosise <katrina.prosise@foundries.io>