Add Reference Manual summary page
Added `reference-manual/index.rst` as a summary/landing page for the
reference manual.

Note not every topic was mentioned; the RM section itself is likely to
be thinned out, as many pages are more appropriate for the user guide.

QA: ran linkcheck, checked html output in browser, ran linter.

This commit applies to:

Jira FFTK-1973

Signed-off-by: Katrina Prosise <katrina.prosise@foundries.io>