rm: security: Add boot flow keys graph and UEFI and TF-A info
Add two more certificates: UEFI and TF-A. While on this, add some
leading text and link to the detailed information for each certificate.

It includes the directories for UEFI and TF-A, and a leading text with a
link on where to find more info for UEFI.

Add the diagram for imx, intel, stm32 boot flow as they differ.

While on this, fix some texts to satisfy vale linter and a leading text
for Secure Connection to Cloud Services section.

Add the command line to generate the certificate file again.

Signed-off-by: Daiane Angolini <daiane.angolini@foundries.io>