Cleanup docker tutorial
The Docker tutorial section had a number of style and grammar fixes
applied. Information was also updated to reflect the usage of the Fast
First Target, so mentions of build wait time was removed, as well as the
separate branches.

QA Steps: Content was reviewed for correctness where possible, however a detailed
follow along was skipped. Linter was ran. HTML output was checked for
correct rendering. No issues were spotted.

This commit addresses FFTK-2511
This commit applies to FFTK-2510
This commit applies to FFTK-2430
This commit applies to FFTK-988

Signed-off-by: Katrina Prosise <katrina.prosise@foundries.io>