Cleanup working with tags tutorial
Tutorial edited for simplification and style guide.
Additional changes were needed to reflect the introduction of fast first
Target in v91. As this change means there are no longer devel and master
branches, much of this tutorial needed to be modified. Information was
added pointing to the reference manual page on having the CI follow
branches. This is a stop-gap measure until an additional tutorial is
added which will focus on adding an additional branch. At that point,
this tutorial will be revised again.

QA steps: ran linter, checked output in browser.

This commit addresses FFTK-2517
This commit applies to FFTK-2510
This commit applies to FFTK-2430
This commit applies to FFTK-988

Signed-off-by: Katrina Prosise <katrina.prosise@foundries.io>