Fix broken links in security section
Linkcheck encountered issues with some links. In two instances it was a
simple 404 issue with a changed location of content. In the other the
link functions in the browser, however linkcheck seemed to have issue
with GitHub links linking to specific line numbers.

For the link in `cert-rotation.rst`, it was simply removed, as it did
not feel necessary.

QA Steps: ran linkcheck

No related issues/tickets, minor quick fix.

Signed-off-by: Katrina Prosise <katrina.prosise@foundries.io>