Cleanup user guide sections a-c
Account management, container preloading, custom CI, and custom SOTA
client sections were all cleaned up in line with the style guide.
Container preloading was also updated to remove references of named
branches (devel), and removed a number of screenshots of the UI which
are covered in the getting started and tutorial sections.

QA Steps: viewed rendered html in browser. Ran vale linter and
linkcheck.

This commit applies to FFTK-988

Signed-off-by: Katrina Prosise <katrina.prosise@foundries.io>