Cleanup reference-manual qemu section
As the QEMU section had been cleaned up as parts of previous commits,
the focus here was to address any outstanding issues.
the qcow2 artifact was reported to no longer be available, so the
instructions were updated to reflect this.

Output was checked in the browser.

This commit addresses FFTK-2793
This commit applies to FFTK-2732
This commit applies to FFTK 2731

Signed-off-by: Katrina Prosise <katrina.prosise@foundries.io>