Cleanup reference manual lmp section
`reference-manual/linux` pages were checked and edited for grammar,
spelling, and style.

Pages which were a better fit for the user-guide were relocated under
lmP Customization; an index page for that section was also added.

Cleanup and possible relocation of linux-updating was skipped to avoid
merge conflicts.

Redirects added for moved pages.

QA steps: checked rendered html. Ran linter within editor. Ran
linkcheck. No issues to report, however this is a relatively large commit.

This commit addresses FFTK 2791
This commit addresses FFTK 1887
This commit addresses FFTK 1741
This commit addresses FFTK 1681
This commit applies to FFTK 1687
This commit applies to FFTK 2730
This commit applies to FFTK 2732
This commit applies to FFTK 2731
This commit applies to FFTK 1702
This commit applies to FFTK 988
This commit applies to FFTK 2733

Signed-off-by: Katrina Prosise <katrina.prosise@foundries.io>