Cleanup ota reference manual
Grammar and style changes made.
Updates to content may be required.

QA: Ran linter in editor.
Checked rendered output in browser, no rendering issues spotted.
Linkcheck encountering issues not believed to be associated with any
changes made in this commit.

This commit addresses FFTK-2792
This commit applies to FFTK-2730
This commit applies to FFTK-2732
This commit applies to FFTK-2731
This commit applies to FFTK-1702
This commit applies to FFTK-988

Signed-off-by: Katrina Prosise <katrina.prosise@foundries.io>