rm: factory: Add info on TUF targets expiration configuration
It turned the current info on the TUF targets expiration
configuration is not quite accurate. It suggests to use
`<build-type>.params.TUF_TARGETS_EXPIRE` parameters which is rather an
internal mechanism and should not be recommended for users.

The proper way to set targets metadata default validity period is
through the `tuf.targets_expire_after` parameter, this is the one that
we should recommend to users.

Signed-off-by: Mike Sul <mike.sul@foundries.io>