Update general and existing glossary terms
The glossary was cleaned up a bit, new terms added, as well as links to
related topic pages.

Links to glossary terms were added in a few areas, with a focus on the
'getting started' sections.

This is the first of a handful of PR commits meant to update the glossary.

QA Steps: Ran build, examined output. No issues spotted. Linkcheck
failed due to unrelated issue which has been identified and will be
updated separately.

This commit addresses FFTK-3008

Signed-off-by: Katrina Prosise <katrina.prosise@foundries.io>