Add meilisearch for docs search
Added search functionality.

Two api keys are exposed, with limited scope.
The search api, which is normal and required for basic search
functionality.
An additional key allows for getting the list of indexes.
This allows for checking if an index matching the release exists.

If no index exists, the default doc search is used.

Known issues: as a limit to the number of indexes/releases will
eventually present itself, older indexes will need to be deleted.
As Sphinx is static, this presents a bit of an issue.
However, once version selection is implemented, older versions will be rebuilt,
meaning older releases with deleted search indexes will fallback to
using the default Sphinx search.

Additionally, local/PR CI builds will use a "dev" index.
This will eventually be configured to use the most recent commit,
however for now the index was created by scrapping "latest".

Note this commit does NOT cover the creation of indexes, which at this
time will be handled manually, and requires private keys.

QA: Checked rendered output, and tested a number of cases by modifying
the variables used.

This commit addresses task FFTK-3215
This commit applies to task FFTK-3091

Signed-off-by: Katrina Prosise <katrina.prosise@foundries.io>