Add offline documentation via pdf
Added basic configuration and support for generating PDF version of the
documentation. This is done with simplepdf.

Also added to the jobserv.yml file for PR requests, which will need to be
tested.

QA steps: built PDF and HTML. Checked that HTML would build with missing
PDF. Checked that PDf download worked from both top index page and any
other page. Checked that PDF links worked and output was readable.
Note that simplepdf will report errors due to links in sidebar.These can
be ignored.

This commit addresses task FFTK-1795
This commit applies to FFPK-93
This commit applies to FS-1826

Signed-off-by: Katrina Prosise <katrina.prosise@foundries.io>