Add offline documentation
Offline documentation is provided in both PDF and HTML form.
As the extension used to build the PDF may come with dependency issues,
it is commented out in `requirements.txt`. Both offline versions will be
created manually as part of the release process by a technical writer
and added to GitHub.

Change to the glossary page fixes an issue with linking
within the PDF.the TOC on the title page shows 2 levels, which makes the
PDF simpler to navigate.

QA: Checked that "Offline Version" link in footer would link to
`/releases/` if dev version of docs, otherwise to the release version.
Checked rendered output in browser. Ran linkcheck.

This commit addresses FFTK-1785

Signed-off-by: Katrina Prosise <katrina.prosise@foundries.io>