Minor cleanup to rest api page
Focus was on changing the api link so that customers not logged in would see
a notice of such rather than a page error. An admonition to this effect was
added.
All links to the api were changed as otherwise they would need individual
references.

Other minor clean done.

No related ticket, quick fix.

Signed-off-by: Katrina Prosise <katrina.prosise@foundries.io>