Update Sphinx
Sphinx and its depends have have been updated to more recent versions.
For Sphinx this brings a number of improvements, and brings it from v5.3
to v8.1.

It should be noted that this will require python 3.11 or greater.
Running linkcheck locally resulted in a number of ssl cert errors. These
errors are a new feature. However it does not appear as if it can be
disabled. It is also possible that the errors resulted from a local
issue.

QA Steps: Ran linkcheck (see above note), checked rendered HTML, checked
that PDF built as _expected_

No related tasks/issues, regular maintenance task

Signed-off-by: Katrina Prosise <katrina.prosise@foundries.io>