Expand acronyms
Reviewed results from vale, and expanded where appropriate.

The Fioctl install section and the Troubleshooting guide were not
addressed, so as to avoid having to deal with a merge conflict from
recently merged PRs, and will be addressed separately.

Ran build and checked output for errors.

This commit applies to issue FFTK-2012, "…prefix acronyms…"

Signed-off-by: Katrina Prosise <katrina.prosise@foundries.io>