Align and reorganize qemu pages
Emulation with QEMU page deleted, as it does not serve much of a
purpose.
Other QEMU pages edited to have one parameter per line for the command
example; this is so it can be easier to visually parse.

QA Steps: Built HTML with no issues or warnings. Checked rendered
output. Edited with linter plugin on.

This commit addresses issue FFTK-4004, "Align QEMU pages"

Signed-off-by: Katrina Prosise <katrina.prosise@foundries.io>